Recepten 90% en discount Klaar voor 70%:
Recepten TODO:
- get products of that id en niet alles

Discount TODO:
- discount CRUD ZELF AANPASSEN  VAN BRAND ID

+++ AppieApplication/AppieApplication/ViewModel/BrandViewModel.cs
@@ -25,7 +25,7 @@ namespace AppieApplication.ViewModel
 
         public int Id { get { return brand.id; } set { brand.id = value; RaisePropertyChanged(); } }
         public String Name { get { return brand.Name; } set { brand.Name = value; RaisePropertyChanged(); } }
-        public int Price { get { return brand.Price; } set { brand.Price = value; RaisePropertyChanged(); } }
+        public double Price { get { return brand.Price; } set { brand.Price = value; RaisePropertyChanged(); } }
 
 
     }
